Adds An Individual Job Endpoint
Per Growth, we've been asked to have the job posting urls in the format
`/jobs/:slug` rather than `/jobs/:shortcode` this means that we litterally have
to loop through every job and check against our generated slug for matches. This
is not scalable but we will have to rethink things if we have 100s of job
listings anyway.

Caching has been pulled back to the model from the router, to allow the single
jobs request to use the cached request for multiples.

These requests are cached in the same manner.

I also fixed a bug in the Redis caching scheme that would try to fall back to a
null value if there was an error. The error is now passed on if the backup is a
null value.